Add: org-indent face
This fontifies org-indent to match the `org-default` face so that indentation characters match the color of the background.  This is important in case the org-indent character is not a space (e.g. when using variable-pitch fonts in Org buffers, using a * for indentation is better than a space, because it aligns entry content with headlines, while spaces result in content that is not indented enough).