Add doom-manegarm theme
Manegarm is a dark, autumn-inspired theme.
Its main feature is the non-standard foreground color, namely green.

I've chosen the name because the color palette, once it was finished,
reminded me a lot of album covers of some bands that I used to listen
to. The one that stuck out the most was the Månegarm album "Vredens Tid".