homage-white: darken selection colour (grey -> base3)
Grey is set intentionally dark in this theme, because a lot of elements
are coloured with grey in base theme and homage themes palette is not
working great with it. Instead of tweaking individual elements in homage
theme files I decided just make grey as a dark colour.

The consequence of it is using grey (i.e. base8) is not working in
selection boxes. Using direct base colour fixes this problem, but
possibly I should fix grey colour and fix individual elemnts instead.